Merged branches/HWRF into the trunk, as approved by the Dev Meeting on the 16th of Nov 2011.
These changes are:

1. Major bug fix: COLLECT should be COLLECT_LOC in the INITPOSTs.
Otherwise, the NMM post will fail when run with more than two
processors.

2. New products: model output surface wind stress X and Y components.
The MDLTAUX and MDLTAUY contain the NMM output TAUX and TAUY variables.
RQSTFIELD fields 601 and 602 will output those wind stress components
instead of calling CALTAU.  This is needed for HWRF since we need to
see the TAUX and TAUY that are fed to the ocean model, rather than the
possibly different values the post will generate.

3. The mp_physics and cu_physics parameters were not being
read in by the INITPOSTs. That is now corrected for
all NMM INITPOSTs.



git-svn-id: https://svn-dtc-unifiedpostproc.cgd.ucar.edu/trunk@117 3ad70301-0a67-45a7-a7d4-3cbace1a8cc5