Added support for producing simulated satellite radiances for channels 15-18 of the SSMIS onboard the F17 satellite. When doing so, the CRTM is instructed to process only the channels 15-18 using the new SELECT_CHANNELS subroutine in SELECT_CHANNELS.f.  For other images, CRTM is still processing all channels the imager has.
With the F17 SSMIS, parameter 118 in table 2 is used, 
which is the generic "brightness temperature."  This is
done since it maps directly to the generic "brightness
temperature" in GRIB2.  Channel numbers are handled
using a level and level type, also to allow easy
translation between GRIB1 and GRIB2

To differentiate between the different channels and
nadir versus non-nadir, below-ground soil levels are
used.  The first level is 117cm for non-nadir corrected
radiances and 167cm for nadir radiances.  The second
level is the channel number in cm (15cm for channel 15,
16cm for channel 16, etc.)  This ends up showing up in grads
as this:

    brtmp117_15cm 0 118 ** 117-15 cm underground Brightness temperature [K]
    brtmp117_16cm 0 118 ** 117-16 cm underground Brightness temperature [K]
    brtmp117_17cm 0 118 ** 117-17 cm underground Brightness temperature [K]
    brtmp117_18cm 0 118 ** 117-18 cm underground Brightness temperature [K]

and is able to translate to the same variable in GRIB2,
unlike our previous arrangement where it showed up as
near-tropopause shear, a much more meaningless
mapping.  



git-svn-id: https://svn-dtc-unifiedpostproc.cgd.ucar.edu/branches/microwave-53deg-non-nadir@182 3ad70301-0a67-45a7-a7d4-3cbace1a8cc5