*** TRUNK COMMIT *** This commit brings the trunk to UPPV2.0 - the merger of EMC and DTC unipost.  From this point forward the unipost/src directory of unipost will be the same as EMC, except the makefile.  The libraries are currently not sync'd, but the effort to attain this has begun.  The EMC NCEPLIBS repository will be copied to the external libraries and linked to unipost as CRTM lib is today.
MERGING NOTE:  Some files have not changed, some files have simple changes, but large
changes exist (including file/diectory renaming) in this commit.  A merge command via
svn or other automated utility based on the previous version should be carefully examined
to insure the operation was successful due to the nature of this change.

Detailed notes on the change are available at:
http://www.dtcenter.org/wrf-nmm/users/overview/upp_updates.php

Direct question to wrfhelp


git-svn-id: https://svn-dtc-unifiedpostproc.cgd.ucar.edu/trunk@311 3ad70301-0a67-45a7-a7d4-3cbace1a8cc5