** TRUNK COMMIT **
WARNING -- this is one of several commits that will leave the trunk in a compromised
state until all commits are complete.  They are done one at a time for the ability to
modularly work with them in the future.

Reconfiguration and updates for w3 library.  Atttained from Eugene Mirvis via
yellowstone.  Breaks w3 into w3emc and w3nco.



git-svn-id: https://svn-dtc-unifiedpostproc.cgd.ucar.edu/trunk@391 3ad70301-0a67-45a7-a7d4-3cbace1a8cc5