== TRUNK COMMIT ==
This fixes the trunk to compile -- an erorr was made in the merge
of my working branch to the trunk and not all changes were brought
over to support the new libraries.  New run scripts are needed 
because of a new directory structure.  W3 was broke into 2 libraries
and this change is required for copygb and ndate makefile.  Some
files were not deleted and are no longer used. 



git-svn-id: https://svn-dtc-unifiedpostproc.cgd.ucar.edu/trunk@395 3ad70301-0a67-45a7-a7d4-3cbace1a8cc5