** BRANCH COMMIT **
Replaying commit 367 and 409 to INITPOST_NMM_BIN_MPIIO.f instead of
INITPOST_NMM_BIN_MPIIO_IJK.f.

Deleting INITPOST_NMM_BIN_MPIIO_IJK.f and INITPOST_NMM_BIN_MPIIO_IKJ.f
as there is no rule to make or use these files anymore.

The original commit message was:
r367 | samuel.trahan@noaa.gov | 2013-04-02 09:13:57 -0600 (Tue, 02 Apr 2013) | 9 lines

 This commit adds two bug fixes:

 1. When the MesoSAS is selected (cu_physics=85), the UPP treats it as
 SAS (sets icu_physics=4)

 2. The synthetic satellite code no longer adds bogus microphysics
 unless the BMJ scheme is in use (icu_physics==2).

r409 | samuel.trahan@noaa.gov | 2013-04-15 11:08:33 -0600 (Mon, 15 Apr 2013) | 36 lines

This commit adds a bugfix to a bug that is as old as the NMM-E model
itself.  The iplib required that the projection center be the same as
the domain center when using the WRF-NMM grid (rotated
latitude-longitude semistaggered Arakawa E, which is GRIB1 grid 203).
This bugfix changes the iplib routine to remove that restriction.  The
result is a somewhat drastic improvement in product quality

Consequentially, the UPP is changed to accomadate this bugfix.  This
update removes Gopal's workaround for the iplib bug: the UPP used to
"lie" about the projection center, claiming that center is at the
domain center.  This update includes a fix for that: UPP now reports
the correct projection information.  The domain center is still needed
by post-processor scripts that decide where to place
latitude-longitude grids when converting from E grid to lat-lon.  For
that purpose, the UPP now generates a new file:
"this-domain-center.ksh" which sets two variables "clat" and "clon" to
the domain center latitude and longitude, respectively.  Any
postprocessor scripts will need to read that file (instead of the
wgrib -V output) to get the domain center location.

However, this means that YOUR DPHI AND DLAMBDA (DX and DY IN WRF) MUST
BE A MULTIPLE OF 0.001 DEGREES in order to use this fix.  Note that dx
and dy are the H-to-V gridpoint distance, not the grid resolution.
The fix will still work with non-conforming dx/dy/dphi/dlambda values,
but the interpolation will be slightly wrong.

It should be noted that this update cannot be used by some other
WRF-NMM models due to their use of either very small or irrational dx
and dy values that are not a multiple of 0.001.  The original iplib
implementation correctly handled such values but only for the single
domain WRF-NMM -- multiple domains were not correctly handled before.
This implementation correctly handles nests with arbitrary starting
locations but only if the dphi and dlambda are correctly represented
in the GRIB file, where they only have a resolution of 0.001 degrees.




git-svn-id: https://svn-dtc-unifiedpostproc.cgd.ucar.edu/branches/HWRF@470 3ad70301-0a67-45a7-a7d4-3cbace1a8cc5