** BRANCH COMMIT **
Adding commit r469 to NMM binary IO. Previous log was:

Critical fix for Central Pacific HWRF: when the nest is exactly centered on
the date line, 50% of the time the HWRF will mistakenly output lon=0 (or
nearly 0) due to averaging the locations of the two points stratling the
baseline (180-180=0). This fix uses a more complicated calcuation in those
cases. Results will not change except for the Central Pacific, in cases in
which the code would have broken before.




git-svn-id: https://svn-dtc-unifiedpostproc.cgd.ucar.edu/branches/HWRF@471 3ad70301-0a67-45a7-a7d4-3cbace1a8cc5