*** BRANCH COMMIT ***
Adding Direct read of reflectivity from thompson microphysics from wrfouts.
This feature will be in v2.3 but since we are currecntly using a modified 
v2.2, these are added.  

Likely only a temporary need as ultimate goal is to put to the AGL reflectivity into
the wrf diags.  UPP product here until that is done.

Mods simply read in REFL_10CM from wrfout and assign to UPP array REF_10CM (new in v2.3)
and uses this to interpolate the reflectivity to 1/4km. No actual derivation/calculation
of reflectivity is done in UPP for this.

  MDL2AGL.f      -- Add imp_physics check for thomspon (8) and use REF_10CM instead of DBZ
  INITPOST.F     -- Add read of REFL_10CM from wrf to REF_10CM for upp
  ALLOCATE_ALL.f -- Allocate REF_10CM
  VRBLS3D_mod.f  -- Add REF_10CM to variable list
  DEALLOCATE.f   -- Deallocate REF_10CM



git-svn-id: https://svn-dtc-unifiedpostproc.cgd.ucar.edu/branches/ncar_ens_2015@648 3ad70301-0a67-45a7-a7d4-3cbace1a8cc5